+++ jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    moduleFileExtensions: [
+        "ts",
+        "js",
+    ],
+    testMatch: [
+        "**/*.test.ts"
+    ],
+};