+++ src/polaris-log-properties.ts
@@ -1,6 +1,8 @@
-import { EventKindDescriptionLogProperty } from './entities/event-kind-description-log-property';
-import { RealityLogProperty } from './entities/reality-log-property';
-import { RequestLogProperty } from './entities/request-log-property';
+import {
+    EventKindDescriptionLogProperty,
+    RealityLogProperty,
+    RequestLogProperty,
+} from './entities';
 
 export interface PolarisLogProperties {
     throwable?: object;